+++ server/index.js
@@ -1,8 +1,11 @@
 const express = require('express');
-const server = express();
 const mysql = require('mysql');
 const dbcredentials = require('./_config');
+const groceryRouter = require('./routes/groceries/index.js');
+
+const server = express();
 server.use(express.json());
+server.use('/api', groceryRouter);
 
 function makeQuery(sql) {
   const connection = mysql.createConnection(dbcredentials.credentials);
@@ -16,64 +19,6 @@ function makeQuery(sql) {
   });
   return sqlPromise;
 }
-
-server.get('/api/groceries', async (req, res) => {
-  const sql =
-    'SELECT groceryItems.id AS id, itemName, remainingAmount, unitName, locationName ' +
-    'FROM groceryItems ' +
-    'JOIN amountUnits ' +
-      'ON amountUnitid = amountUnits.id ' +
-    'JOIN storageLocations ' +
-      'ON locationId = storageLocations.id';
-  const results = await makeQuery(sql)
-    .catch(() => res.status(500).send('An error occurred while connecting to the database'));
-  res.status(200).send(results);
-});
-
-server.get('/api/groceries/:id', async (req, res) => {
-  if (!req.params.id) {
-    res.status(400).send('Item ID required');
-  } else {
-    const sql =
-      'SELECT * ' +
-      'FROM groceryItems ' +
-      'JOIN groceryCategories ' +
-        'ON categoryId = groceryCategories.id ' +
-      'JOIN amountUnits ' +
-        'ON amountUnitid = amountUnits.id ' +
-      'JOIN storageLocations ' +
-        'ON locationId = storageLocations.id ' +
-      `WHERE groceryItems.id = ${mysql.escape(req.params.id)}`;
-    const results = await makeQuery(sql)
-      .catch(() => res.status(500).send('An error occurred while connecting to the database'));
-    if (!results.length) {
-      res.status(404).send(`Item with ID ${req.params.id} not found`);
-    } else {
-      res.status(200).send({
-        'id': results[0].id,
-        'name': results[0].itemName,
-        'category': {
-          'id': results[0].categoryId,
-          'name': results[0].categoryName
-        },
-        'amount': {
-          'initial': results[0].amount,
-          'quantity': results[0].remainingAmount,
-          'unit': results[0].unitName,
-          'unitId': results[0].amountUnitId
-        },
-        'purchaseDate': results[0].purchaseDate.toISOString().substring(0, 10),
-        'expirationDate': results[0].expirationDate.toISOString().substring(0, 10),
-        'location': {
-          'id': results[0].locationId,
-          'name': results[0].locationName,
-          'description': results[0].description
-        },
-        'notes': results[0].notes
-      });
-    }
-  }
-});
 
 server.get('/api/locations', async (req, res) => {
   const sql =
@@ -136,21 +81,6 @@ server.get('/api/units', async (req, res) => {
   res.status(200).send(results);
 });
 
-server.delete('/api/groceries/:id', async (req, res) => {
-  if (!req.params.id) {
-    res.status(400).send('Student ID required');
-  } else {
-    const sql = `DELETE FROM groceryItems WHERE id = ${mysql.escape(req.params.id)}`;
-    const results = await makeQuery(sql)
-      .catch(() => { res.status(500).send('An Error occurred while connecting to the database'); });
-    if (!results.affectedRows) {
-      res.status(404).send(`Item with ID ${req.params.id} not found`);
-    } else {
-      res.status(200).send({ 'id': Number(req.params.id) });
-    }
-  }
-});
-
 server.delete('/api/locations/:id', async (req, res) => {
   if (!req.params.id) {
     res.status(400).send('Student ID required');
@@ -173,32 +103,6 @@ server.delete('/api/locations/:id', async (req, res) => {
   }
 });
 
-server.patch('/api/groceries', async (req, res) => {
-  const reqProps = ['id', 'name', 'category', 'amount', 'amountRemaining', 'unit', 'purchaseDate', 'expirationDate', 'location', 'notes'];
-  for (const prop in reqProps) {
-    if (!req.body.hasOwnProperty(reqProps[prop])) {
-      res.status(400).send(`Item ${reqProps[prop]} required.`);
-      return;
-    }
-  }
-  const {
-    id, name, category, amount, amountRemaining, unit, purchaseDate, expirationDate, location, notes
-  } = req.body;
-  let sql =
-    'UPDATE groceryItems ' +
-    'SET itemName = ?, categoryId = ?, amount = ?, remainingAmount = ?, amountUnitId = ?, purchaseDate = ?, expirationDate = ?, locationId = ?, notes = ? ' +
-    'WHERE id = ?';
-  const inserts = [name, category, amount, amountRemaining, unit, purchaseDate, expirationDate, location, notes, id];
-  sql = mysql.format(sql, inserts);
-  const results = await makeQuery(sql)
-    .catch(() => { res.status(500).send('An Error occurred while connecting to the database'); });
-  if (!results.affectedRows) {
-    res.status(404).send(`Item with ID ${req.params.id} not found`);
-  } else {
-    res.status(200).send({ 'id': Number(req.body.id) });
-  }
-});
-
 server.patch('/api/locations', async (req, res) => {
   const reqProps = ['id', 'name', 'description'];
   for (const prop in reqProps) {
@@ -218,28 +122,6 @@ server.patch('/api/locations', async (req, res) => {
   } else {
     res.status(200).send({ 'id': Number(req.body.id) });
   }
-});
-
-server.post('/api/groceries', async (req, res) => {
-  const reqProps = ['name', 'category', 'amount', 'amountRemaining', 'unit', 'purchaseDate', 'expirationDate', 'location', 'notes'];
-  for (const prop in reqProps) {
-    if (!req.body.hasOwnProperty(reqProps[prop])) {
-      res.status(400).send(`Item ${reqProps[prop]} required.`);
-      return;
-    }
-  }
-  const {
-    name, category, amount, amountRemaining, unit, purchaseDate, expirationDate, location, notes
-  } = req.body;
-  let sql =
-    'INSERT INTO groceryItems ' +
-    '(id, itemName, categoryId, amount, remainingAmount, amountUnitId, purchaseDate, expirationDate, locationId, notes) ' +
-    'VALUES (NULL, ?, ?, ?, ?, ?, ?, ?, ?, ?)';
-  const inserts = [name, category, amount, amountRemaining, unit, purchaseDate, expirationDate, location, notes];
-  sql = mysql.format(sql, inserts);
-  const results = await makeQuery(sql)
-    .catch(() => { res.status(500).send('An Error occurred while connecting to the database'); });
-  res.status(200).send({ 'id': results.insertId });
 });
 
 server.post('/api/locations', async (req, res) => {
